+++ pilot/pkg/serviceregistry/kube/controller/serviceexportcache.go
@@ -104,7 +104,7 @@ type discoverabilityPolicySelector func(*model.Service) model.EndpointDiscoverab
 type serviceExportCacheImpl struct {
 	*Controller
 
-	serviceExports client.Erased
+	serviceExports client.Untyped
 
 	// clusterLocalPolicySelector selects an appropriate EndpointDiscoverabilityPolicy for the cluster.local host.
 	clusterLocalPolicySelector discoverabilityPolicySelector
@@ -203,7 +203,7 @@ func (ec *serviceExportCacheImpl) Run(stop <-chan struct{}) {
 		return
 	}
 	dInformer := ec.client.DynamicInformer().ForResource(mcs.ServiceExportGVR).Informer()
-	ec.serviceExports = client.NewErased(ec.client, dInformer, client.Filter{ObjectFilter: ec.opts.GetFilter()})
+	ec.serviceExports = client.NewUntyped(ec.client, dInformer, client.Filter{ObjectFilter: ec.opts.GetFilter()})
 	// Register callbacks for events.
 	registerHandlers[controllers.Object](ec.Controller, ec.serviceExports, "ServiceExports", ec.onServiceExportEvent, nil)
 	go dInformer.Run(stop)
