--- pkg/kube/client/erased.go
@@ -1,69 +1,0 @@
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package client
-
-import (
-	klabels "k8s.io/apimachinery/pkg/labels"
-	"k8s.io/client-go/tools/cache"
-
-	"istio.io/istio/pilot/pkg/util/informermetric"
-	"istio.io/istio/pkg/kube"
-	"istio.io/istio/pkg/kube/controllers"
-	"istio.io/istio/pkg/ptr"
-	"istio.io/pkg/log"
-)
-
-// Erased is a Cached with
-type Erased interface {
-	// Get looks up an object by name and namespace. If it does not exist, nil is returned
-	Get(name, namespace string) controllers.Object
-	// List looks up an object by namespace and labels.
-	// Use metav1.NamespaceAll and klabels.Everything() to select everything.
-	List(namespace string, selector klabels.Selector) []controllers.Object
-	// ListUnfiltered is like List but ignores any *client side* filters previously configured.
-	ListUnfiltered(namespace string, selector klabels.Selector) []controllers.Object
-
-	// AddEventHandler inserts a handler. The handler will be called for all Create/Update/Removals.
-	// When ShutdownHandlers is called, the handler is removed.
-	AddEventHandler(h cache.ResourceEventHandler)
-	// HasSynced returns true when the informer is initially populated and that all handlers added
-	// via AddEventHandler have been called with the initial state.
-	// note: this differs from a standard informer HasSynced, which does not check handlers have been called.
-	HasSynced() bool
-	// ShutdownHandlers terminates all handlers added by AddEventHandler.
-	// Warning: this only applies to handlers called via AddEventHandler; any handlers directly added
-	// to the underlying informer are not touched
-	ShutdownHandlers()
-}
-
-func NewErased(c kube.Client, inf cache.SharedIndexInformer, filter Filter) Erased {
-	return ptr.Of(newReadClient[controllers.Object](c, inf, filter))
-}
-
-func newReadClient[T controllers.Object](c kube.Client, inf cache.SharedIndexInformer, filter Filter) readClient[T] {
-	if filter.ObjectTransform != nil {
-		_ = inf.SetTransform(filter.ObjectTransform)
-	} else {
-		_ = inf.SetTransform(kube.StripUnusedFields)
-	}
-
-	if err := inf.SetWatchErrorHandler(informermetric.ErrorHandlerForCluster(c.ClusterID())); err != nil {
-		log.Debugf("failed to set watch handler, informer may already be started: %v", err)
-	}
-	return readClient[T]{
-		inf:    inf,
-		filter: filter.ObjectFilter,
-	}
-}
